+++ app/src/main/java/digital/lamp/mindlamp/sensor/WifiData.kt
@@ -2,8 +2,8 @@ package digital.lamp.mindlamp.sensor
 
 import android.content.ContentValues
 import android.content.Context
-import com.mindlamp.Lamp
-import com.mindlamp.WiFi
+import digital.lamp.Lamp
+import digital.lamp.WiFi
 import digital.lamp.mindlamp.R
 import digital.lamp.mindlamp.network.model.DimensionData
 import digital.lamp.mindlamp.network.model.LogEventRequest
