+++ sensor-core/src/main/java/digital/lamp/Accelerometer.java
@@ -1,5 +1,5 @@
 
-package com.mindlamp;
+package digital.lamp;
 
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
@@ -16,66 +16,49 @@ import android.os.IBinder;
 import android.os.PowerManager;
 import android.provider.BaseColumns;
 import android.util.Log;
-
-import com.mindlamp.utils.LampConstants;
-import com.mindlamp.utils.Lamp_Sensor;
-
+import digital.lamp.utils.LampConstants;
+import digital.lamp.utils.Lamp_Sensor;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * LAMP Light module
- * - Light raw data
- * - Light sensor information
+ * LAMP Accelerometer module
+ * - Accelerometer raw data
+ * - Accelerometer sensor information
  *
  * @author df
  */
-public class Light extends Lamp_Sensor implements SensorEventListener {
+public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
 
-    /**
-     * Logging tag (default = "LAMP::Light")
-     */
-    private static String TAG = "LAMP::Light";
+    public static String TAG = "LAMP::Accelerometer";
 
     private static SensorManager mSensorManager;
-    private static Sensor mLight;
+    private static Sensor mAccelerometer;
 
     private static HandlerThread sensorThread = null;
     private static Handler sensorHandler = null;
-
     private static PowerManager.WakeLock wakeLock = null;
+    private static String LABEL = "";
 
-    private static Float LAST_VALUE = null;
+    private static Float[] LAST_VALUES = null;
     private static long LAST_TS = 0;
     private static long LAST_SAVE = 0;
 
     private static int FREQUENCY = -1;
     private static double THRESHOLD = 0;
-    // Reject any data points that come in more often than frequency
-    private static boolean ENFORCE_FREQUENCY = false;
 
-    /**
-     * Broadcasted event: new light values
-     * ContentProvider: LightProvider
-     */
-    public static final String ACTION_LAMP_LIGHT = "ACTION_LAMP_LIGHT";
-    public static final String ACTION_LAMP_LIGHT_LABEL = "ACTION_LAMP_LIGHT_LABEL";
+    public static final String ACTION_LAMP_ACCELEROMETER = "ACTION_LAMP_ACCELEROMETER";
+    public static final String ACTION_LAMP_ACCELEROMETER_LABEL = "ACTION_LAMP_ACCELEROMETER_LABEL";
     public static final String EXTRA_LABEL = "label";
 
-    /**
-     * Until today, no available Android phone samples higher than 208Hz (Nexus 7).
-     * http://ilessendata.blogspot.com/2012/11/android-accelerometer-sampling-rates.html
-     */
-    private List<ContentValues> data_values = new ArrayList<ContentValues>();
-
-    private static String LABEL = "";
+    private List<ContentValues> data_values = new ArrayList<>();
 
     private static DataLabel dataLabeler = new DataLabel();
 
     public static class DataLabel extends BroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent intent) {
-            if (intent.getAction().equals(ACTION_LAMP_LIGHT_LABEL)) {
+            if (intent.getAction().equals(ACTION_LAMP_ACCELEROMETER_LABEL)) {
                 LABEL = intent.getStringExtra(EXTRA_LABEL);
             }
         }
@@ -89,22 +72,25 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
     @Override
     public void onSensorChanged(SensorEvent event) {
         long TS = System.currentTimeMillis();
-        if (ENFORCE_FREQUENCY && TS < LAST_TS + FREQUENCY / 1000)
+        if ((TS - LAST_TS) < LampConstants.INTERVAL)
             return;
-        if (LAST_VALUE != null && THRESHOLD > 0 && Math.abs(event.values[0] - LAST_VALUE) < THRESHOLD) {
+        if (LAST_VALUES != null && THRESHOLD > 0 && Math.abs(event.values[0] - LAST_VALUES[0]) < THRESHOLD
+                && Math.abs(event.values[1] - LAST_VALUES[1]) < THRESHOLD
+                && Math.abs(event.values[2] - LAST_VALUES[2]) < THRESHOLD) {
             return;
         }
 
-        LAST_TS = TS;
-        LAST_VALUE = event.values[0];
+        LAST_VALUES = new Float[]{event.values[0], event.values[1], event.values[2]};
 
         ContentValues rowData = new ContentValues();
-        rowData.put(Light_Data.TIMESTAMP, TS);
-        rowData.put(Light_Data.LIGHT_LUX, event.values[0]);
-        rowData.put(Light_Data.ACCURACY, event.accuracy);
-        rowData.put(Light_Data.LABEL, LABEL);
+        rowData.put(Accelerometer_Data.TIMESTAMP, TS);
+        rowData.put(Accelerometer_Data.VALUES_0, event.values[0]);
+        rowData.put(Accelerometer_Data.VALUES_1, event.values[1]);
+        rowData.put(Accelerometer_Data.VALUES_2, event.values[2]);
+        rowData.put(Accelerometer_Data.ACCURACY, event.accuracy);
+        rowData.put(Accelerometer_Data.LABEL, LABEL);
 
-        if (awareSensor != null) awareSensor.onLightChanged(rowData);
+        if (awareSensor != null) awareSensor.onAccelerometerChanged(rowData);
 
         data_values.add(rowData);
         LAST_TS = TS;
@@ -120,28 +106,24 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
         LAST_SAVE = TS;
     }
 
-    private static Light.LAMPSensorObserver awareSensor;
-
-    public static void setSensorObserver(Light.LAMPSensorObserver observer) {
+    private static LAMPSensorObserver awareSensor;
+    public static void setSensorObserver(LAMPSensorObserver observer) {
         awareSensor = observer;
     }
-
-    public static Light.LAMPSensorObserver getSensorObserver() {
+    public static LAMPSensorObserver getSensorObserver() {
         return awareSensor;
     }
 
     public interface LAMPSensorObserver {
-        void onLightChanged(ContentValues data);
+        void onAccelerometerChanged(ContentValues data);
     }
 
     @Override
     public void onCreate() {
         super.onCreate();
 
-        AUTHORITY = getPackageName() + ".provider.light";
-
         mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
-        mLight = mSensorManager.getDefaultSensor(Sensor.TYPE_LIGHT);
+        mAccelerometer = mSensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
 
         sensorThread = new HandlerThread(TAG);
         sensorThread.start();
@@ -153,11 +135,10 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
         sensorHandler = new Handler(sensorThread.getLooper());
 
         IntentFilter filter = new IntentFilter();
-        filter.addAction(ACTION_LAMP_LIGHT_LABEL);
+        filter.addAction(ACTION_LAMP_ACCELEROMETER_LABEL);
         registerReceiver(dataLabeler, filter);
 
-
-        if (Lamp.DEBUG) Log.d(TAG, "Light service created!");
+        if (Lamp.DEBUG) Log.d(TAG, "Accelerometer service created!");
     }
 
     @Override
@@ -165,14 +146,11 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
         super.onDestroy();
 
         sensorHandler.removeCallbacksAndMessages(null);
-        mSensorManager.unregisterListener(this, mLight);
+        mSensorManager.unregisterListener(this, mAccelerometer);
         sensorThread.quit();
-
         wakeLock.release();
 
         unregisterReceiver(dataLabeler);
-
-        if (Lamp.DEBUG) Log.d(TAG, "Light service terminated...");
     }
 
     @Override
@@ -180,28 +158,27 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
         super.onStartCommand(intent, flags, startId);
 
         if (PERMISSIONS_OK) {
-            if (mLight == null) {
-
+            if (mAccelerometer == null) {
                 stopSelf();
             } else {
-
-                int new_frequency = LampConstants.FREQUENCY_LIGHT;
-                double new_threshold = LampConstants.THRESHOLD_LIGHT;
+                int new_frequency = LampConstants.FREQUENCY_ACCELEROMETER;
+                double new_threshold = LampConstants.THRESHOLD_ACCELEROMETER;
 
                 if (FREQUENCY != new_frequency
                         || THRESHOLD != new_threshold) {
 
                     sensorHandler.removeCallbacksAndMessages(null);
-                    mSensorManager.unregisterListener(this, mLight);
+                    mSensorManager.unregisterListener(this, mAccelerometer);
 
                     FREQUENCY = new_frequency;
                     THRESHOLD = new_threshold;
                 }
 
-                mSensorManager.registerListener(this, mLight, new_frequency, sensorHandler);
+                mSensorManager.registerListener(this, mAccelerometer, FREQUENCY, sensorHandler);
+                LAST_SAVE = System.currentTimeMillis();
 
+                if (Lamp.DEBUG) Log.d(TAG, "Accelerometer service active: " + FREQUENCY + " ms");
 
-                if (Lamp.DEBUG) Log.d(TAG, "Light service active: " + FREQUENCY + "ms");
             }
         }
 
@@ -213,14 +190,15 @@ public class Light extends Lamp_Sensor implements SensorEventListener {
         return null;
     }
 
-    public static final class Light_Data implements BaseColumns {
+    public static final class Accelerometer_Data implements BaseColumns {
 
         public static final String _ID = "_id";
         public static final String TIMESTAMP = "timestamp";
         public static final String DEVICE_ID = "device_id";
-        public static final String LIGHT_LUX = "double_light_lux";
+        public static final String VALUES_0 = "double_values_0";
+        public static final String VALUES_1 = "double_values_1";
+        public static final String VALUES_2 = "double_values_2";
         public static final String ACCURACY = "accuracy";
         public static final String LABEL = "label";
     }
-
 }
