+++ sensor-core/src/main/java/digital/lamp/Magnetometer.java
@@ -1,5 +1,5 @@
 
-package com.mindlamp;
+package digital.lamp;
 
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
@@ -16,28 +16,28 @@ import android.os.IBinder;
 import android.os.PowerManager;
 import android.provider.BaseColumns;
 import android.util.Log;
-import com.mindlamp.utils.LampConstants;
-import com.mindlamp.utils.Lamp_Sensor;
+import digital.lamp.utils.LampConstants;
+import digital.lamp.utils.Lamp_Sensor;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * LAMP Rotation module
- * - Rotation raw data
- * - Rotation sensor information
+ * LAMP Magnetometer module
+ * - Magnetometer raw data
+ * - Magnetometer sensor information
  *
  * @author df
  */
-public class Rotation extends Lamp_Sensor implements SensorEventListener {
+public class Magnetometer extends Lamp_Sensor implements SensorEventListener {
 
     /**
-     * Logging tag (default = "LAMP::Rotation")
+     * Logging tag (default = "LAMP::Magnetometer")
      */
-    private static String TAG = "LAMP::Rotation";
+    private static String TAG = "LAMP::Magnetometer";
 
     private static SensorManager mSensorManager;
-    private static Sensor mRotation;
+    private static Sensor mMagnetometer;
 
     private static HandlerThread sensorThread = null;
     private static Handler sensorHandler = null;
@@ -52,18 +52,18 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
     // Reject any data points that come in more often than frequency
 
     /**
-     * Broadcasted event: new rotation values
-     * ContentProvider: RotationProvider
+     * Broadcasted event: new sensor values
+     * ContentProvider: MagnetometerProvider
      */
-    public static final String ACTION_LAMP_ROTATION = "ACTION_LAMP_ROTATION";
-    public static final String ACTION_LAMP_ROTATION_LABEL = "ACTION_LAMP_ROTATION_LABEL";
+    public static final String ACTION_LAMP_MAGNETOMETER = "ACTION_LAMP_MAGNETOMETER";
+    public static final String ACTION_LAMP_MAGNETOMETER_LABEL = "ACTION_LAMP_MAGNETOMETER_LABEL";
     public static final String EXTRA_LABEL = "label";
 
     /**
      * Until today, no available Android phone samples higher than 208Hz (Nexus 7).
      * http://ilessendata.blogspot.com/2012/11/android-accelerometer-sampling-rates.html
      */
-    private List<ContentValues> data_values = new ArrayList<>();
+    private List<ContentValues> data_values = new ArrayList<ContentValues>();
 
     private static String LABEL = "";
 
@@ -72,7 +72,7 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
     public static class DataLabel extends BroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent intent) {
-            if (intent.getAction().equals(ACTION_LAMP_ROTATION_LABEL)) {
+            if (intent.getAction().equals(ACTION_LAMP_MAGNETOMETER_LABEL)) {
                 LABEL = intent.getStringExtra(EXTRA_LABEL);
             }
         }
@@ -85,39 +85,28 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
 
     @Override
     public void onSensorChanged(SensorEvent event) {
-        if (SignificantMotion.isSignificantMotionActive && !SignificantMotion.CURRENT_SIGMOTION_STATE) {
-            if (data_values.size() > 0) {
-                final ContentValues[] data_buffer = new ContentValues[data_values.size()];
-                data_values.toArray(data_buffer);
-
-                data_values.clear();
-            }
-            return;
-        }
-
         long TS = System.currentTimeMillis();
         if ((TS - LAST_TS) < LampConstants.INTERVAL)
             return;
-        if (LAST_VALUES != null && THRESHOLD > 0 && Math.abs(event.values[0] - LAST_VALUES[0]) < THRESHOLD
-                && Math.abs(event.values[1] - LAST_VALUES[1]) < THRESHOLD
-                && Math.abs(event.values[2] - LAST_VALUES[2]) < THRESHOLD) {
+        if (LAST_VALUES != null && THRESHOLD > 0 &&
+                Math.abs(event.values[0] - LAST_VALUES[0]) < THRESHOLD &&
+                Math.abs(event.values[0] - LAST_VALUES[1]) < THRESHOLD &&
+                Math.abs(event.values[0] - LAST_VALUES[2]) < THRESHOLD) {
+            float val = Math.abs(event.values[0] - LAST_VALUES[0]);
             return;
         }
 
         LAST_VALUES = new Float[]{event.values[0], event.values[1], event.values[2]};
 
         ContentValues rowData = new ContentValues();
-        rowData.put(Rotation_Data.TIMESTAMP, TS);
-        rowData.put(Rotation_Data.VALUES_0, event.values[0]);
-        rowData.put(Rotation_Data.VALUES_1, event.values[1]);
-        rowData.put(Rotation_Data.VALUES_2, event.values[2]);
-        if (event.values.length == 4) {
-            rowData.put(Rotation_Data.VALUES_3, event.values[3]);
-        }
-        rowData.put(Rotation_Data.ACCURACY, event.accuracy);
-        rowData.put(Rotation_Data.LABEL, LABEL);
+        rowData.put(Magnetometer_Data.TIMESTAMP, TS);
+        rowData.put(Magnetometer_Data.VALUES_0, event.values[0]);
+        rowData.put(Magnetometer_Data.VALUES_1, event.values[1]);
+        rowData.put(Magnetometer_Data.VALUES_2, event.values[2]);
+        rowData.put(Magnetometer_Data.ACCURACY, event.accuracy);
+        rowData.put(Magnetometer_Data.LABEL, LABEL);
 
-        if (awareSensor != null) awareSensor.onRotationChanged(rowData);
+        if (awareSensor != null) awareSensor.onMagnetometerChanged(rowData);
 
         data_values.add(rowData);
         LAST_TS = TS;
@@ -133,30 +122,29 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         LAST_SAVE = TS;
     }
 
-    private static Rotation.LAMPSensorObserver awareSensor;
+    private static Magnetometer.LAMPSensorObserver awareSensor;
 
-    public static void setSensorObserver(Rotation.LAMPSensorObserver observer) {
+    public static void setSensorObserver(Magnetometer.LAMPSensorObserver observer) {
         awareSensor = observer;
     }
 
-    public static Rotation.LAMPSensorObserver getSensorObserver() {
+    public static Magnetometer.LAMPSensorObserver getSensorObserver() {
         return awareSensor;
     }
 
     public interface LAMPSensorObserver {
-        void onRotationChanged(ContentValues data);
+        void onMagnetometerChanged(ContentValues data);
     }
+
 
     @Override
     public void onCreate() {
         super.onCreate();
 
-//        AUTHORITY = Rotation_Provider.getAuthority(this);
-
-        TAG = "Aware::Rotation";
+//        AUTHORITY = Magnetometer_Provider.getAuthority(this);
 
         mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
-        mRotation = mSensorManager.getDefaultSensor(Sensor.TYPE_ROTATION_VECTOR);
+        mMagnetometer = mSensorManager.getDefaultSensor(Sensor.TYPE_MAGNETIC_FIELD);
 
         sensorThread = new HandlerThread(TAG);
         sensorThread.start();
@@ -168,10 +156,10 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         sensorHandler = new Handler(sensorThread.getLooper());
 
         IntentFilter filter = new IntentFilter();
-        filter.addAction(ACTION_LAMP_ROTATION_LABEL);
+        filter.addAction(ACTION_LAMP_MAGNETOMETER_LABEL);
         registerReceiver(dataLabeler, filter);
 
-        if (Lamp.DEBUG) Log.d(TAG, "Rotation service created!");
+        if (Lamp.DEBUG) Log.d(TAG, "Magnetometer service created!");
     }
 
     @Override
@@ -179,14 +167,14 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         super.onDestroy();
 
         sensorHandler.removeCallbacksAndMessages(null);
-        mSensorManager.unregisterListener(this, mRotation);
+        mSensorManager.unregisterListener(this, mMagnetometer);
         sensorThread.quit();
 
         wakeLock.release();
 
         unregisterReceiver(dataLabeler);
 
-        if (Lamp.DEBUG) Log.d(TAG, "Rotation service terminated...");
+        if (Lamp.DEBUG) Log.d(TAG, "Magnetometer service terminated...");
     }
 
     @Override
@@ -194,28 +182,27 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         super.onStartCommand(intent, flags, startId);
 
         if (PERMISSIONS_OK) {
-            if (mRotation == null) {
+            if (mMagnetometer == null) {
                 stopSelf();
             } else {
+                int new_frequency = LampConstants.FREQUENCY_MAGNETOMETER;
+                double new_threshold = LampConstants.THRESHOLD_MAGNETOMETER;
 
-                int new_frequency = LampConstants.FREQUENCY_ROTATION;
-                double new_threshold = LampConstants.THRESHOLD_ROTATION;
-                              if (FREQUENCY != new_frequency
+                if (FREQUENCY != new_frequency
                         || THRESHOLD != new_threshold) {
 
                     sensorHandler.removeCallbacksAndMessages(null);
-                    mSensorManager.unregisterListener(this, mRotation);
+                    mSensorManager.unregisterListener(this, mMagnetometer);
 
                     FREQUENCY = new_frequency;
                     THRESHOLD = new_threshold;
                 }
 
-                mSensorManager.registerListener(this, mRotation, new_frequency,sensorHandler);
+                mSensorManager.registerListener(this, mMagnetometer, new_frequency, sensorHandler);
                 LAST_SAVE = System.currentTimeMillis();
 
-                if (Lamp.DEBUG) Log.d(TAG, "Rotation service active...");
+                if (Lamp.DEBUG) Log.d(TAG, "Magnetometer service active...");
             }
-
         }
 
         return START_STICKY;
@@ -226,7 +213,7 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         return null;
     }
 
-    public static final class Rotation_Data implements BaseColumns {
+    public static final class Magnetometer_Data implements BaseColumns {
 
         public static final String _ID = "_id";
         public static final String TIMESTAMP = "timestamp";
@@ -234,7 +221,6 @@ public class Rotation extends Lamp_Sensor implements SensorEventListener {
         public static final String VALUES_0 = "double_values_0";
         public static final String VALUES_1 = "double_values_1";
         public static final String VALUES_2 = "double_values_2";
-        public static final String VALUES_3 = "double_values_3";
         public static final String ACCURACY = "accuracy";
         public static final String LABEL = "label";
     }
