+++ sensor-core/src/main/java/digital/lamp/Rotation.java
@@ -1,5 +1,5 @@
 
-package com.mindlamp;
+package digital.lamp;
 
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
@@ -16,29 +16,32 @@ import android.os.IBinder;
 import android.os.PowerManager;
 import android.provider.BaseColumns;
 import android.util.Log;
-import com.mindlamp.utils.LampConstants;
-import com.mindlamp.utils.Lamp_Sensor;
+import digital.lamp.utils.LampConstants;
+import digital.lamp.utils.Lamp_Sensor;
+
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * LAMP Accelerometer module
- * - Accelerometer raw data
- * - Accelerometer sensor information
+ * LAMP Rotation module
+ * - Rotation raw data
+ * - Rotation sensor information
  *
  * @author df
  */
-public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
+public class Rotation extends Lamp_Sensor implements SensorEventListener {
 
-    public static String TAG = "LAMP::Accelerometer";
+    /**
+     * Logging tag (default = "LAMP::Rotation")
+     */
+    private static String TAG = "LAMP::Rotation";
 
     private static SensorManager mSensorManager;
-    private static Sensor mAccelerometer;
+    private static Sensor mRotation;
 
     private static HandlerThread sensorThread = null;
     private static Handler sensorHandler = null;
     private static PowerManager.WakeLock wakeLock = null;
-    private static String LABEL = "";
 
     private static Float[] LAST_VALUES = null;
     private static long LAST_TS = 0;
@@ -46,19 +49,30 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
 
     private static int FREQUENCY = -1;
     private static double THRESHOLD = 0;
+    // Reject any data points that come in more often than frequency
 
-    public static final String ACTION_LAMP_ACCELEROMETER = "ACTION_LAMP_ACCELEROMETER";
-    public static final String ACTION_LAMP_ACCELEROMETER_LABEL = "ACTION_LAMP_ACCELEROMETER_LABEL";
+    /**
+     * Broadcasted event: new rotation values
+     * ContentProvider: RotationProvider
+     */
+    public static final String ACTION_LAMP_ROTATION = "ACTION_LAMP_ROTATION";
+    public static final String ACTION_LAMP_ROTATION_LABEL = "ACTION_LAMP_ROTATION_LABEL";
     public static final String EXTRA_LABEL = "label";
 
+    /**
+     * Until today, no available Android phone samples higher than 208Hz (Nexus 7).
+     * http://ilessendata.blogspot.com/2012/11/android-accelerometer-sampling-rates.html
+     */
     private List<ContentValues> data_values = new ArrayList<>();
+
+    private static String LABEL = "";
 
     private static DataLabel dataLabeler = new DataLabel();
 
     public static class DataLabel extends BroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent intent) {
-            if (intent.getAction().equals(ACTION_LAMP_ACCELEROMETER_LABEL)) {
+            if (intent.getAction().equals(ACTION_LAMP_ROTATION_LABEL)) {
                 LABEL = intent.getStringExtra(EXTRA_LABEL);
             }
         }
@@ -71,6 +85,16 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
 
     @Override
     public void onSensorChanged(SensorEvent event) {
+        if (SignificantMotion.isSignificantMotionActive && !SignificantMotion.CURRENT_SIGMOTION_STATE) {
+            if (data_values.size() > 0) {
+                final ContentValues[] data_buffer = new ContentValues[data_values.size()];
+                data_values.toArray(data_buffer);
+
+                data_values.clear();
+            }
+            return;
+        }
+
         long TS = System.currentTimeMillis();
         if ((TS - LAST_TS) < LampConstants.INTERVAL)
             return;
@@ -83,14 +107,17 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         LAST_VALUES = new Float[]{event.values[0], event.values[1], event.values[2]};
 
         ContentValues rowData = new ContentValues();
-        rowData.put(Accelerometer_Data.TIMESTAMP, TS);
-        rowData.put(Accelerometer_Data.VALUES_0, event.values[0]);
-        rowData.put(Accelerometer_Data.VALUES_1, event.values[1]);
-        rowData.put(Accelerometer_Data.VALUES_2, event.values[2]);
-        rowData.put(Accelerometer_Data.ACCURACY, event.accuracy);
-        rowData.put(Accelerometer_Data.LABEL, LABEL);
+        rowData.put(Rotation_Data.TIMESTAMP, TS);
+        rowData.put(Rotation_Data.VALUES_0, event.values[0]);
+        rowData.put(Rotation_Data.VALUES_1, event.values[1]);
+        rowData.put(Rotation_Data.VALUES_2, event.values[2]);
+        if (event.values.length == 4) {
+            rowData.put(Rotation_Data.VALUES_3, event.values[3]);
+        }
+        rowData.put(Rotation_Data.ACCURACY, event.accuracy);
+        rowData.put(Rotation_Data.LABEL, LABEL);
 
-        if (awareSensor != null) awareSensor.onAccelerometerChanged(rowData);
+        if (awareSensor != null) awareSensor.onRotationChanged(rowData);
 
         data_values.add(rowData);
         LAST_TS = TS;
@@ -106,24 +133,30 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         LAST_SAVE = TS;
     }
 
-    private static LAMPSensorObserver awareSensor;
-    public static void setSensorObserver(LAMPSensorObserver observer) {
+    private static Rotation.LAMPSensorObserver awareSensor;
+
+    public static void setSensorObserver(Rotation.LAMPSensorObserver observer) {
         awareSensor = observer;
     }
-    public static LAMPSensorObserver getSensorObserver() {
+
+    public static Rotation.LAMPSensorObserver getSensorObserver() {
         return awareSensor;
     }
 
     public interface LAMPSensorObserver {
-        void onAccelerometerChanged(ContentValues data);
+        void onRotationChanged(ContentValues data);
     }
 
     @Override
     public void onCreate() {
         super.onCreate();
 
+//        AUTHORITY = Rotation_Provider.getAuthority(this);
+
+        TAG = "Aware::Rotation";
+
         mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
-        mAccelerometer = mSensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
+        mRotation = mSensorManager.getDefaultSensor(Sensor.TYPE_ROTATION_VECTOR);
 
         sensorThread = new HandlerThread(TAG);
         sensorThread.start();
@@ -135,10 +168,10 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         sensorHandler = new Handler(sensorThread.getLooper());
 
         IntentFilter filter = new IntentFilter();
-        filter.addAction(ACTION_LAMP_ACCELEROMETER_LABEL);
+        filter.addAction(ACTION_LAMP_ROTATION_LABEL);
         registerReceiver(dataLabeler, filter);
 
-        if (Lamp.DEBUG) Log.d(TAG, "Accelerometer service created!");
+        if (Lamp.DEBUG) Log.d(TAG, "Rotation service created!");
     }
 
     @Override
@@ -146,11 +179,14 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         super.onDestroy();
 
         sensorHandler.removeCallbacksAndMessages(null);
-        mSensorManager.unregisterListener(this, mAccelerometer);
+        mSensorManager.unregisterListener(this, mRotation);
         sensorThread.quit();
+
         wakeLock.release();
 
         unregisterReceiver(dataLabeler);
+
+        if (Lamp.DEBUG) Log.d(TAG, "Rotation service terminated...");
     }
 
     @Override
@@ -158,28 +194,28 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         super.onStartCommand(intent, flags, startId);
 
         if (PERMISSIONS_OK) {
-            if (mAccelerometer == null) {
+            if (mRotation == null) {
                 stopSelf();
             } else {
-                int new_frequency = LampConstants.FREQUENCY_ACCELEROMETER;
-                double new_threshold = LampConstants.THRESHOLD_ACCELEROMETER;
 
-                if (FREQUENCY != new_frequency
+                int new_frequency = LampConstants.FREQUENCY_ROTATION;
+                double new_threshold = LampConstants.THRESHOLD_ROTATION;
+                              if (FREQUENCY != new_frequency
                         || THRESHOLD != new_threshold) {
 
                     sensorHandler.removeCallbacksAndMessages(null);
-                    mSensorManager.unregisterListener(this, mAccelerometer);
+                    mSensorManager.unregisterListener(this, mRotation);
 
                     FREQUENCY = new_frequency;
                     THRESHOLD = new_threshold;
                 }
 
-                mSensorManager.registerListener(this, mAccelerometer, FREQUENCY, sensorHandler);
+                mSensorManager.registerListener(this, mRotation, new_frequency,sensorHandler);
                 LAST_SAVE = System.currentTimeMillis();
 
-                if (Lamp.DEBUG) Log.d(TAG, "Accelerometer service active: " + FREQUENCY + " ms");
-
+                if (Lamp.DEBUG) Log.d(TAG, "Rotation service active...");
             }
+
         }
 
         return START_STICKY;
@@ -190,7 +226,7 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         return null;
     }
 
-    public static final class Accelerometer_Data implements BaseColumns {
+    public static final class Rotation_Data implements BaseColumns {
 
         public static final String _ID = "_id";
         public static final String TIMESTAMP = "timestamp";
@@ -198,6 +234,7 @@ public class Accelerometer extends Lamp_Sensor implements SensorEventListener {
         public static final String VALUES_0 = "double_values_0";
         public static final String VALUES_1 = "double_values_1";
         public static final String VALUES_2 = "double_values_2";
+        public static final String VALUES_3 = "double_values_3";
         public static final String ACCURACY = "accuracy";
         public static final String LABEL = "label";
     }
