+++ sensor-core/src/main/java/digital/lamp/providers/Significant_Provider.java
@@ -1,5 +1,5 @@
 
-package com.mindlamp.providers;
+package digital.lamp.providers;
 
 
 import android.content.Context;
