+++ sensor-core/src/main/java/digital/lamp/ui/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package com.mindlamp.ui;
+package digital.lamp.ui;
 
 import android.content.Context;
 import android.content.SharedPreferences;
